% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Module 1: Obtain and Clean the NOAA earthquake data}
\usage{
eq_clean_data(datatoclean)
}
\arguments{
\item{datatoclean}{A data frame with raw data obtained from NOAA website}
}
\value{
A data frame with cleaned date, latitude and longitude numerical columns
}
\description{
Module 1: Obtain and Clean the NOAA earthquake data
}
\details{
After downloading and reading in the dataset from NOAA site,
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}.
National Geophysical Data Center / World Data Service (NGDC/WDS): Significant Earthquake Database.
National Geophysical Data Center
The function returns a date column created by uniting the year, month, day and
converting it to the Date class and also converts LATITUDE and LONGITUDE columns
to numeric class.
}
\examples{
\dontrun{
data <- readr::read_delim("signif.txt", delim = "\t")
data <- eq_clean_data(data)
}

}
