% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleaning NOAA earthquake data location values}
\usage{
eq_location_clean(locationtoclean)
}
\arguments{
\item{locationtoclean}{A data frame with raw data obtained from NOAA website}
}
\value{
A data frame with cleaned LOCATION_NAME column
}
\description{
Cleaning NOAA earthquake data location values
}
\details{
The function cleans the LOCATION_NAME column
}
\examples{
\dontrun{
library(readr)
data <- readr::read_delim("signif.txt", delim = "\t")
data <- eq_location_clean(data)
}

}
