% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Module 2: Visualize the times at which earthquakes occur within certain countries}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Create asthetics mapping.}

\item{data}{The data}

\item{stat}{The statistical transformation as a string.}

\item{position}{Position adjustment}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{show the legends or not}

\item{inherit.aes}{the default aesthetics}

\item{...}{Other arguments passed on to [layer()]. These are
  often aesthetics, used to set an aesthetic to a fixed value, like
  `colour = "red"` or `size = 3`. They may also be parameters
  to the paired geom/stat.

#' @section Aesthetics:
\code{geom_timeline} understands the following aesthetics

\itemize{
  \item \strong{\code{x}} # Time variable
  \item \code{y}          # Factor indicating some stratification
  \item \code{color}      # Color of border of elements
  \item \code{shape}      # Shape
  \item \code{size}       # Size
  \item \code{alpha}      # Transparency (1: opaque; 0: transparent)
  \item \code{fill}       # Color of inside of elements
  \item \code{stroke}     # Stroke
}}
}
\value{
Produce a time line plot of selected earthquakes.
}
\description{
Module 2: Visualize the times at which earthquakes occur within certain countries
}
\details{
The data to be used with this geom from NOAA website
}
\examples{
# The data must be cleaned using the function \code{eq_clean_data}, included in the package.
# Aesthetics can be specified in the \code{ggplot} function or in \code{geom_timeline} geom function
\dontrun{
data <- readr::read_delim("signif.txt", delim = "\t")
data <- eq_clean_data(data)
data \%>\%
dplyr::filter(COUNTRY == c("MEXICO","USA") & lubridate::year(DATE) >= 2010) \%>\%
ggplot(aes(x=DATE,y=COUNTRY,color=TOTAL_DEATHS,size=EQ_PRIMARY)) +
geom_timeline(alpha=.5) +
theme(legend.position="bottom", legend.box="horizontal", plot.title=element_text(hjust=0.5)) +
ggtitle("Earthquakes Visualization Tool") +
labs(size = "Richter scale value", color = "# deaths")
}
}
