% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Module 2: Visualization Tools: Build a geom for ggplot2 called geom_timeline()
for plotting a time line of earthquakes
ranging from xmin to xmaxdates with a point for each earthquake.}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{The mappings created.}

\item{data}{The data need to display in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{Position adjustment}

\item{na.rm}{Missing values are removed with if false}

\item{show.legend}{logical. This layer be included in the legends or not}

\item{inherit.aes}{Override the default aesthetic if false}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}.

\code{geom_timeline} understands the following aesthetics

\itemize{
\item \code{x}          # Time variable (required)
\item \code{y}          # Factor indicating some stratification
\item \code{color}      # Color of border of elements
\item \code{shape}      # Shape
\item \code{size}       # Size
\item \code{alpha}      # Transparency (1: opaque; 0: transparent)
\item \code{fill}       # Color of inside of elements
\item \code{stroke}     # Stroke
}}
}
\value{
The geom \code{geom_timeline} used with the \code{ggplot} function
to create a time line plot of selected earthquakes.
}
\description{
The xaesthetic is a date and an optional y aesthetic is a factor indicating
some stratification in which case multiple time lines will be plotted for
each level of the factor (e.g. country).
}
\details{
The data is downloaded and readed from NOAA website,
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}.
}
\examples{
The data need to be cleaned using the function \code{eq_clean_data}
\dontrun{
data <- readr::read_delim("signif.txt", delim = "\t")
data <- eq_clean_data(data)
data \%>\%
dplyr::filter(COUNTRY == c("CANADA","USA") & lubridate::year(DATE) >= 2010) \%>\%
ggplot(aes(x=DATE,y=COUNTRY,color=TOTAL_DEATHS,size=EQ_PRIMARY)) +
geom_timeline(alpha=.5) +
theme(legend.position="bottom", legend.box="horizontal", plot.title=element_text(hjust=0.5)) +
ggtitle("Earthquakes Visualization Tool") +
labs(size = "Richter scale value", color = "# deaths")
}
}
